backend/requirements/local.txt
-r base.txt

# Testing
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
factory-boy>=3.3,<4.0
pytest-cov>=5.0,<6.0

# Code quality
ruff>=0.9,<1.0

# Development tools
ipython>=8.0,<9.0
django-extensions>=3.2,<4.0
